refactor: Let fonts draw their own text, free Text of SpriteBatch
