+++ CutTheRopeDX/Framework/Visual/FontStashFont.cs
@@ -1,9 +1,13 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
+
+using CutTheRopeDX.Desktop;
 
 using FontStashSharp;
 
 using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
 
 namespace CutTheRopeDX.Framework.Visual
 {
@@ -195,106 +199,313 @@ namespace CutTheRopeDX.Framework.Visual
             // The actual rendering is done differently with FontStashSharp
             return null;
         }
-    }
 
-    /// <summary>
-    /// Configuration for font effects (stroke, shadow).
-    /// </summary>
-    internal sealed class FontEffectSettings
-    {
-        /// <summary>
-        /// Whether stroke is enabled.
-        /// </summary>
-        public bool HasStroke { get; set; }
+        /// <inheritdoc />
+        public override bool DrawsOwnText => true;
 
         /// <summary>
-        /// Stroke thickness in pixels.
+        /// Rasterizer state with scissor test enabled for text clipping.
         /// </summary>
-        public int StrokeAmount { get; set; } = 1;
-
-        /// <summary>
-        /// Stroke color.
-        /// </summary>
-        public Color StrokeColor { get; set; } = Color.Black;
-
-        /// <summary>
-        /// Whether shadow is enabled.
-        /// </summary>
-        public bool HasShadow { get; set; }
-
-        /// <summary>
-        /// Shadow horizontal offset in pixels.
-        /// </summary>
-        public int ShadowOffsetX { get; set; }
-
-        /// <summary>
-        /// Shadow vertical offset in pixels.
-        /// </summary>
-        public int ShadowOffsetY { get; set; }
-
-        /// <summary>
-        /// Shadow color.
-        /// </summary>
-        public Color ShadowColor { get; set; } = Color.Black;
-
-        /// <summary>
-        /// Returns a settings instance with no effects.
-        /// </summary>
-        public static FontEffectSettings None => new();
-
-        /// <summary>
-        /// Creates settings with stroke only.
-        /// </summary>
-        /// <param name="amount">Stroke thickness in pixels.</param>
-        /// <param name="color">Stroke color, defaults to black.</param>
-        /// <returns>A <see cref="FontEffectSettings"/> instance configured with stroke only.</returns>
-        public static FontEffectSettings CreateStroke(int amount = 1, Color? color = null)
+        private static readonly RasterizerState ScissorRasterizerState = new()
         {
-            return new FontEffectSettings
-            {
-                HasStroke = true,
-                StrokeAmount = amount,
-                StrokeColor = color ?? Color.Black
-            };
+            CullMode = CullMode.None,
+            ScissorTestEnable = true
+        };
+
+        /// <summary>
+        /// Cached render targets for compositing text layers (shadow, stroke, fill) at full
+        /// opacity before applying the fade alpha uniformly. Consecutive text draws alternate
+        /// targets so a target is not rewritten immediately after being sampled.
+        /// </summary>
+        private static readonly RenderTarget2D[] s_textCompositeTargets = new RenderTarget2D[2];
+
+        /// <summary>
+        /// Index of the render target used by the previous composite text draw.
+        /// </summary>
+        private static int s_textCompositeTargetIndex = -1;
+
+        /// <summary>
+        /// Selects the other composite render target for the next text draw.
+        /// </summary>
+        /// <param name="currentIndex">The target index used by the previous draw.</param>
+        /// <returns>The target index to use for the next draw.</returns>
+        internal static int GetNextCompositeTargetIndex(int currentIndex)
+        {
+            return (currentIndex + 1) % s_textCompositeTargets.Length;
         }
 
         /// <summary>
-        /// Creates settings with shadow only.
+        /// Renders text using FontStashSharp with stroke, shadow, and color modulation.
+        /// When fading, all layers are first composited at full opacity onto a render target,
+        /// then drawn to screen with the fade alpha so shadow/stroke/fill fade in sync.
         /// </summary>
-        /// <param name="offsetX">Shadow horizontal offset.</param>
-        /// <param name="offsetY">Shadow vertical offset.</param>
-        /// <param name="color">Shadow color, defaults to black.</param>
-        /// <returns>A <see cref="FontEffectSettings"/> instance configured with shadow only.</returns>
-        public static FontEffectSettings CreateShadow(int offsetX, int offsetY, Color? color = null)
+        /// <param name="call">Layout, formatted lines, color modulation, and ping-pong clip state for the text element.</param>
+        public override void DrawText(in TextDrawCall call)
         {
-            return new FontEffectSettings
-            {
-                HasShadow = true,
-                ShadowOffsetX = offsetX,
-                ShadowOffsetY = offsetY,
-                ShadowColor = color ?? Color.Black
-            };
-        }
+            Color parentColor = call.InheritedColor.ToXNA();
 
-        /// <summary>
-        /// Creates settings with both stroke and shadow using black color.
-        /// </summary>
-        /// <param name="strokeAmount">Stroke thickness in pixels.</param>
-        /// <param name="shadowX">Shadow horizontal offset.</param>
-        /// <param name="shadowY">Shadow vertical offset.</param>
-        /// <returns>A <see cref="FontEffectSettings"/> instance configured with both stroke and shadow.</returns>
-        public static FontEffectSettings CreateStrokeAndShadow(int strokeAmount, int shadowX, int shadowY)
-        {
-            return new FontEffectSettings
+            SpriteBatch spriteBatch = Renderer.GetSpriteBatch();
+            if (spriteBatch == null)
             {
-                HasStroke = true,
-                StrokeAmount = strokeAmount,
-                StrokeColor = Color.Black,
-                HasShadow = true,
-                ShadowOffsetX = shadowX,
-                ShadowOffsetY = shadowY,
-                ShadowColor = Color.Black
-            };
+                Debug.WriteLine("FontStash: SpriteBatch is null");
+                return;
+            }
+
+            DynamicSpriteFont internalFont = GetInternalFont();
+            if (internalFont == null)
+            {
+                Debug.WriteLine("FontStash: Internal font is null");
+                return;
+            }
+
+            if (call.Lines == null || call.Lines.Count == 0)
+            {
+                Debug.WriteLine("FontStash: No formatted strings for text");
+                return;
+            }
+
+            FontEffectSettings effects = GetEffectSettings();
+            Color textColor = GetColor();
+
+            // Apply element and inherited color modulation (RGBAColor uses 0-1 floats; textColor uses 0-255 bytes)
+            static byte ScaleByte(byte channel, float factor)
+            {
+                float scaled = channel * factor; // factor already 0-1, so no /255
+                if (scaled < 0f)
+                {
+                    scaled = 0f;
+                }
+                if (scaled > 255f)
+                {
+                    scaled = 255f;
+                }
+                return (byte)scaled;
+            }
+
+            static Color MakeColor(Color baseColor, float redScale, float greenScale, float blueScale, float alphaScale)
+            {
+                byte finalAlpha = (byte)MathHelper.Clamp(baseColor.A / 255f * alphaScale * 255f, 0f, 255f);
+
+                return Color.FromNonPremultiplied(
+                    ScaleByte(baseColor.R, redScale),
+                    ScaleByte(baseColor.G, greenScale),
+                    ScaleByte(baseColor.B, blueScale),
+                    finalAlpha
+                );
+            }
+
+            float inheritedRed = MathHelper.Clamp(parentColor.R / 255f, 0f, 1f);
+            float inheritedGreen = MathHelper.Clamp(parentColor.G / 255f, 0f, 1f);
+            float inheritedBlue = MathHelper.Clamp(parentColor.B / 255f, 0f, 1f);
+            float inheritedAlpha = MathHelper.Clamp(call.ElementColor.AlphaChannel * (parentColor.A / 255f), 0f, 1f);
+
+            bool hasEffects = effects?.HasStroke == true || effects?.HasShadow == true;
+            bool needsComposite = hasEffects && inheritedAlpha < 1f;
+
+            // Build colors: when compositing via render target, draw layers at full opacity;
+            // the fade alpha is applied once when blitting the RT to screen.
+            float layerAlpha = needsComposite ? 1f : inheritedAlpha;
+
+            Color finalColor = MakeColor(textColor, inheritedRed, inheritedGreen, inheritedBlue, layerAlpha);
+
+            float yPos = call.DrawY + GetTopSpacing();
+            int lineHeight = (int)(internalFont.LineHeight + GetLineOffset());
+
+            GraphicsDevice graphicsDevice = Global.GraphicsDevice;
+            // Queued sprite quads must render before text draws above them or changes render targets.
+            Renderer.FlushQuads();
+            Viewport viewport = graphicsDevice.Viewport;
+
+            float viewportScaleX = viewport.Width / SCREEN_WIDTH;
+            float viewportScaleY = viewport.Height / SCREEN_HEIGHT;
+
+            Matrix transformMatrix =
+                Renderer.GetModelViewMatrix() *
+                Matrix.CreateScale(viewportScaleX, viewportScaleY, 1f);
+
+            // When fading multi-layer text, composite all layers at full opacity onto a
+            // render target, then blit with the fade alpha so every layer fades uniformly.
+            RenderTargetBinding[] previousTargets = null;
+            RenderTarget2D textCompositeTarget = null;
+            if (needsComposite)
+            {
+                int rtW = viewport.Width;
+                int rtH = viewport.Height;
+                s_textCompositeTargetIndex = GetNextCompositeTargetIndex(s_textCompositeTargetIndex);
+                textCompositeTarget = s_textCompositeTargets[s_textCompositeTargetIndex];
+                if (textCompositeTarget == null || textCompositeTarget.IsDisposed ||
+                    textCompositeTarget.Width != rtW || textCompositeTarget.Height != rtH)
+                {
+                    textCompositeTarget?.Dispose();
+                    textCompositeTarget = new RenderTarget2D(graphicsDevice, rtW, rtH, false,
+                        SurfaceFormat.Color, DepthFormat.None, 0, RenderTargetUsage.PreserveContents);
+                    s_textCompositeTargets[s_textCompositeTargetIndex] = textCompositeTarget;
+                }
+
+                previousTargets = graphicsDevice.GetRenderTargets();
+                graphicsDevice.SetRenderTarget(textCompositeTarget);
+                graphicsDevice.Clear(Color.Transparent);
+            }
+
+            // Ping-pong clipping: set a scissor rect so overflowing text is clipped
+            bool isPingPonging = call.IsPingPonging;
+            Rectangle previousScissor = graphicsDevice.ScissorRectangle;
+
+            spriteBatch.Begin(
+                SpriteSortMode.Immediate,
+                BlendState.AlphaBlend,
+                SamplerState.LinearClamp,
+                null,
+                ScissorRasterizerState,
+                null,
+                transformMatrix
+            );
+
+            if (isPingPonging)
+            {
+                float clipW = call.PingPongClipWidth;
+                float clipH = call.PingPongClipHeight;
+                // Clip to the parent element's bounds (e.g., button background image)
+                float clipX = call.PingPongClipLeft;
+                float clipY = call.DrawY;
+                // Transform clip rect corners through the full transform matrix (model-view + viewport scale)
+                Vector2 topLeft = Vector2.Transform(new Vector2(clipX, clipY), transformMatrix);
+                Vector2 bottomRight = Vector2.Transform(new Vector2(clipX + clipW, clipY + clipH), transformMatrix);
+                int sx = (int)topLeft.X;
+                int sy = (int)topLeft.Y;
+                int sw = (int)(bottomRight.X - topLeft.X);
+                int sh = (int)(bottomRight.Y - topLeft.Y);
+                graphicsDevice.ScissorRectangle = new Rectangle(sx, sy, sw, sh);
+            }
+
+            // Render each formatted line
+            foreach (FormattedString formattedString in call.Lines)
+            {
+                if (call.MaxHeight != -1f && yPos >= call.DrawY + call.MaxHeight)
+                {
+                    break;
+                }
+
+                float xPos = call.DrawX;
+
+                if (call.Align == 2) // Center
+                {
+                    xPos += (call.WrapWidth - formattedString.width) / 2f;
+                }
+                else if (call.Align == 3) // Right
+                {
+                    xPos += call.WrapWidth - formattedString.width;
+                }
+
+                // When ping-ponging, left-align the text at the clip area's left edge and scroll
+                if (isPingPonging)
+                {
+                    float clipLeft = call.PingPongClipLeft;
+                    xPos = clipLeft - call.PingPongOffset;
+                }
+
+                Vector2 position = new(xPos, yPos);
+
+                // Draw shadow if enabled
+                if (effects?.HasShadow == true)
+                {
+                    Vector2 shadowBasePos = position + new Vector2(effects.ShadowOffsetX, effects.ShadowOffsetY);
+                    int shadowStrokeAmount = effects.HasStroke ? effects.StrokeAmount : 1;
+                    Color shadowColor = MakeColor(
+                        effects.ShadowColor, inheritedRed, inheritedGreen, inheritedBlue, layerAlpha);
+
+                    for (int x = -shadowStrokeAmount; x <= shadowStrokeAmount; x++)
+                    {
+                        for (int y = -shadowStrokeAmount; y <= shadowStrokeAmount; y++)
+                        {
+                            _ = internalFont.DrawText(
+                                spriteBatch,
+                                formattedString.string_,
+                                shadowBasePos + new Vector2(x, y),
+                                shadowColor
+                            );
+                        }
+                    }
+                }
+
+                // Draw stroke if enabled
+                if (effects?.HasStroke == true)
+                {
+                    Color strokeColor = MakeColor(
+                        effects.StrokeColor, inheritedRed, inheritedGreen, inheritedBlue, layerAlpha);
+                    int strokeAmount = effects.StrokeAmount;
+
+                    for (int x = -strokeAmount; x <= strokeAmount; x++)
+                    {
+                        for (int y = -strokeAmount; y <= strokeAmount; y++)
+                        {
+                            if (x != 0 || y != 0)
+                            {
+                                _ = internalFont.DrawText(
+                                    spriteBatch,
+                                    formattedString.string_,
+                                    position + new Vector2(x, y),
+                                    strokeColor
+                                );
+                            }
+                        }
+                    }
+                }
+
+                // Draw main text
+                _ = internalFont.DrawText(
+                    spriteBatch,
+                    formattedString.string_,
+                    position,
+                    finalColor
+                );
+
+                yPos += lineHeight;
+            }
+
+            spriteBatch.End();
+            BlendParams.InvalidateDeviceCache();
+
+            if (isPingPonging)
+            {
+                graphicsDevice.ScissorRectangle = previousScissor;
+            }
+
+            // Blit the composite render target to screen with the uniform fade alpha
+            if (needsComposite)
+            {
+                if (previousTargets != null && previousTargets.Length > 0)
+                {
+                    graphicsDevice.SetRenderTargets(previousTargets);
+                }
+                else
+                {
+                    graphicsDevice.SetRenderTarget(null);
+                }
+
+                byte fadeByte = (byte)MathHelper.Clamp(inheritedAlpha * 255f, 0f, 255f);
+                Color blitColor = new(fadeByte, fadeByte, fadeByte, fadeByte); // premultiplied tint
+
+                spriteBatch.Begin(
+                    SpriteSortMode.Immediate,
+                    BlendState.AlphaBlend,
+                    SamplerState.LinearClamp,
+                    null,
+                    null,
+                    null,
+                    null
+                );
+                spriteBatch.Draw(textCompositeTarget, Vector2.Zero, blitColor);
+                spriteBatch.End();
+                BlendParams.InvalidateDeviceCache();
+
+                // SpriteBatch leaves its texture in slot zero. Mark it unbound so the next
+                // composite pass cannot retain a sampled binding for a writable target.
+                if (ReferenceEquals(graphicsDevice.Textures[0], textCompositeTarget))
+                {
+                    graphicsDevice.Textures[0] = null;
+                }
+            }
         }
     }
 }
